requirements.txt
fastapi
uvicorn[standard]
asyncpg
pyjwt
pydantic[email]
python-jose
python-dotenv
passlib==1.7.4
python-multipart
faker
argon2-cffi
redis
slowapi
